currency_prediction_system/requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Time Series Analysis
statsmodels>=0.13.0
yfinance>=0.1.70

# Deep Learning
tensorflow>=2.10.0
torch>=1.12.0
transformers>=4.20.0

# Ensemble Models
xgboost>=1.6.0

# Data Collection APIs
pyfredapi>=0.4.0
requests>=2.28.0

# NLP and Sentiment Analysis
nltk>=3.7
vaderSentiment>=3.3.2
afinn>=0.1

# Web Framework and Dashboard
flask>=2.2.0
dash>=2.6.0
dash-bootstrap-components>=1.3.0
plotly>=5.10.0

# Database
sqlite3

# Utilities
joblib>=1.2.0
pickle5>=0.0.11
python-dotenv>=0.19.0

# Development
jupyter>=1.0.0
ipykernel>=6.15.0 